backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn==0.30.6


# Database & ORM
sqlalchemy==2.0.36      
alembic==1.18.3
psycopg2-binary==2.9.10

# Authentication & Security
python-jose==3.3.0
argon2-cffi==23.1.0
email_validator==1.3.1


# Data Validation
pydantic==2.9.2
python-multipart==0.0.9


# Environment Variables
python-dotenv==1.0.1


# Utilities
requests==2.32.3

# Async jobs & email
arq
redis
aiosmtplib
fastapi_mail

# Testing
pytest==8.3.2
httpx==0.27.2          # async test client

# Templates
jinja2==3.1.3
